requirements.txt
python-dotenv==1.0.1
pyTelegramBotAPI==4.20.0
peewee~=3.17.5
pip~=24.1.1
certifi~=2024.6.2
requests~=2.32.3
urllib3~=2.2.2
idna~=3.7
setuptools~=70.2.0
schedule~=1.2.2
django-scheduler~=0.10.1